Add dtype inference where it makes sense
- dtype must be specified to atop, due to difficulty of inference
- infer dtype in elemwise and map_blocks

+++ dask/array/tests/test_array_core.py
@@ -2114,7 +2114,7 @@ def test_tril_triu_errors():
 
 def test_atop_names():
     x = da.ones(5, chunks=(2,))
-    y = atop(add, 'i', x, 'i')
+    y = atop(add, 'i', x, 'i', dtype=x.dtype)
     assert y.name.startswith('add')
